% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rYWAASB.R
\docType{data}
\name{data_ge}
\alias{data_ge}
\title{Dataset3: a tibble containing ENV, GEN, REP factors and
GY(grain yield) and HM agronomic traits from the \code{metan} package.}
\format{
A \code{data.frame} with 420 rows in 5 columns.
\describe{
\item{\code{ENV}}{a character vector}
\item{\code{GEN}}{a character vector}
\item{\code{REP}}{a character vector}
\item{\code{GY}}{a numeric vector}
\item{\code{HM}}{a numeric vector}
}
}
\usage{
data(data_ge)
}
\description{
Dataset3: a tibble containing ENV, GEN, REP factors and
GY(grain yield) and HM agronomic traits from the \code{metan} package.
}
\examples{
\donttest{
library(rYWAASB)
data(data_ge)
}
}
\references{
Olivoto, T., & Lúcio, A.D.C.2020. metan: An R package for
multi‐environment trial analysis. Methods in Ecology and Evolution
, 11(6), 783-789.
}
\keyword{datasets}
