% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rYWAASB.R
\docType{data}
\name{dm}
\alias{dm}
\title{Dataset2: a tibble containing GEN, Trait (days to maturity), \code{WAASB} and \code{WAASBY} indexes.}
\format{
A \code{data.frame} with 13 observations (chickpea genotypes)
within rows and columns including the trait i.e. days to maturity (named as \code{Y}),
\code{WAASB} and \code{WAASBY} indexes values.
\describe{
\item{\code{GEN}}{a character vector saved as factor}
\item{\code{Y}}{a numeric vector}
\item{\code{WAASB}}{a numeric vector}
\item{\code{WAASBY}}{a numeric vector}
}
}
\usage{
data(dm)
}
\description{
Dataset2: a tibble containing GEN, Trait (days to maturity), \code{WAASB} and \code{WAASBY} indexes.
}
\examples{
\donttest{
library(rYWAASB)
data(dm)
ranki(dm)
bar_plot1(dm)
bar_plot2(dm)
PCA_biplot(dm)
}
}
\keyword{datasets}
