% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rYWAASB.R
\docType{package}
\name{rYWAASB-package}
\alias{rYWAASB}
\alias{rYWAASB-package}
\title{Simultaneous Selection by Trait and WAASB Index}
\description{
\pkg{rYWAASB} performs a new ranking algorithm based on a
"Y*WAASB" biplot generated by the 'metan' package which is
used in MET(Multi-Environmental Trials).
This package effectively distinguishes the top-ranked genotypes
based on a given trait (e.g. grain yield or any other trait
in agricultural experiments) and the "WAASB" index in the
Genotype-by-environment interaction effect studies.
Note: Fortunately, this package can impute missing
observations and computes them, eliminating any concerns
about their presence in the data set. A complete guide
may be found at: \url{https://github.com/abeyran/rYWAASB/issues}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abeyran/rYWAASB}
  \item Report bugs at \url{https://github.com/abeyran/rYWAASB/issues}
}

}
\author{
{
Ali Arminian \href{mailto:abeyran@gmail.com}{abeyran@gmail.com}
}
}
