% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot1.R
\docType{package}
\name{bar_plot1}
\alias{bar_plot1}
\title{Draw the first barplot of ranks of genotypes}
\usage{
bar_plot1(datap)
}
\arguments{
\item{datap}{The data set}
}
\description{
Proposes a new ranking algorithm that utilizes a "Y*WAASB" biplot generated from the 'metan' package. The aim of the package is to effectively distinguish the top-ranked genotypes. For a detailed explanation of the process of obtaining a "Y*WAASB" biplot and the related data, please refer to the manual included in this package as well as the study by Olivoto & Lúcio (2020) \doi{10.1111/2041-210X.13384}. In this context, "WAASB" refers to the "Weighted Average of Absolute Scores" provided by Olivoto et al. (2019) \doi{10.2134/agronj2019.03.0220}, which quantifies the stability of genotypes across different environments using linear mixed-effect models. In order to run the package, it is necessary to extract the "WAASB" coefficients using the 'metan' package.

Creates the bar plot for a new index for simultaneous selection of genotypes for trait and WAASB index.
}
\details{
{The \code{bar_plot} function creates the barplot of rank of genotypes}
}
\examples{
data(maize)
bar_plot1(maize)

}
\references{
Olivoto, T., et al. (2019) Mean Performance and Stability in Multi-Environment Trials I: Combining Features of AMMI and BLUP Techniques. Agronomy Journal, 111, 1-12. https://doi.org/10.2134/agronj2019.03.0220

H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abeyran/Stability1}
  \item Report bugs at \url{https://github.com/abeyran/Stability1/issues}
}

}
\author{
Ali Arminian (Ph.D.) Plant Geneticist and Breeder

Maintainer: Ali Arminian <abeyran@gmail.com>
}
\keyword{package}
