% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot1.R
\name{bar_plot1}
\alias{bar_plot1}
\title{The first barplot of the ranks of genotypes}
\usage{
bar_plot1(datap)
}
\arguments{
\item{datap}{The data set}
}
\value{
Returns an object of class \verb{gg, ggmatrix}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
\itemize{
\item \code{bar_plot1()} creates a bar plot for the new index (rYWAASB
for individuals) for simultaneous selection of genotypes by
trait and WAASB index using ggplot2.
}
}
\examples{
\donttest{
data(maize)
bar_plot1(maize)
}
}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York, 2016.
}
\author{
{
Ali Arminian \href{mailto:abeyran@gmail.com}{abeyran@gmail.com}
}
}
