% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot2.R
\name{bar_plot2}
\alias{bar_plot2}
\title{The second barplot of the ranks of genotypes}
\usage{
bar_plot2(datap, verbose = FALSE)
}
\arguments{
\item{datap}{The data set}

\item{verbose}{If \code{verbose = TRUE} then some results are printed}
}
\value{
Returns an object of class \verb{gg, graphics}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
\itemize{
\item \code{bar_plot2()} creates the 2nd barplot of the ranks of genotypes
using the \code{graphics} package.
}
}
\examples{
\donttest{
data(maize)
bar_plot2(maize, verbose = FALSE)
}
}
\author{
{
Ali Arminian \href{mailto:abeyran@gmail.com}{abeyran@gmail.com}
}
}
