% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rYWAASB.R
\docType{data}
\name{data_ge}
\alias{data_ge}
\title{Dataset2: a tibble containing ENV, GEN, REP variables and GY(grain yield) and HM traits from the metan package.}
\format{
A \code{data.frame} with 420 rows in and 5 columns.
\describe{
\item{\code{ENV}}{a character vector}
\item{\code{GEN}}{a character vector}
\item{\code{REP}}{a character vector}
\item{\code{GY}}{a numeric vector}
\item{\code{HM}}{a numeric vector}
}
}
\usage{
data(data_ge)
}
\description{
Dataset2: a tibble containing ENV, GEN, REP variables and GY(grain yield) and HM traits from the metan package.
}
\examples{
library(rYWAASB)
data(data_ge)

}
\keyword{datasets}
