% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rYWAASB.R
\docType{data}
\name{maize}
\alias{maize}
\title{Dataset1: a tibble containing GEN, Trait and WAASB index columns.}
\format{
A \code{data.frame} with 20 observations (genotypes) in rows in and 2 columns of the trait and WAASB index values.
\describe{
\item{\code{GEN}}{a character vector}
\item{\code{Y}}{a numeric vector}
\item{\code{WAASB}}{a numeric vector}
}
}
\usage{
data(maize)
}
\description{
Dataset1: a tibble containing GEN, Trait and WAASB index columns.
}
\examples{
library(rYWAASB)
data(maize)

# ranking genotypes:
ranki(maize)

# first plot by ggplot2:
bar_plot1(maize)

# second plot by R:
bar_plot2(maize)

}
\keyword{datasets}
